feat: Use a more unique concurrency group name
If you're working on multiple stacks simultaneously through a workflow matrix strategy and using a single GitHub environment for all stacks, you might encounter the error `Canceling since a higher priority waiting request for 'prod' exists`. 

While this commit helps to avoid the error, it's not foolproof. A more robust solution could be to allow users to specify their own concurrency group in the calling workflow. This feature might be considered for future releases.